Refactor some of gatsby-node.js
