+++ src/gatsby/README.md
@@ -1,0 +1,3 @@
+# Gatsby support files
+
+These are files that are used by the `gatsby-*.js` config (eg, `gatsby-node.js`).
