+++ src/gatsby/on_create_webpack_config.js
@@ -1,0 +1,44 @@
+// @flow
+
+const { resolve } = require('path')
+
+const root = (...args) => resolve(__dirname, '..', '..', ...args)
+
+/**
+ * Modify Webpack configuration.
+ *
+ * This makes it so that jQuery isn't part of the final JS bundle, saving up
+ * some space.
+ */
+
+const onCreateWebpackConfig = ({ loaders, actions } /*: any */) => {
+  const noop = root('src/lib/helpers/noop.js')
+
+  actions.setWebpackConfig({
+    // Be sure our internals are babelified; unfortunately I can't make this work right now.
+    // module: {
+    //   rules: [
+    //     {
+    //       test: /\.jsx?$/,
+    //       use: [
+    //         loaders.js({
+    //           exclude: /node_modules/,
+    //           include: [/node_modules\/@devhints/, /node_modules\/styled-jsx/]
+    //         })
+    //       ]
+    //     }
+    //   ]
+    // },
+
+    // isotope-layout tries to require('jquery'), but let's let that
+    // fail silently. We don't want it to load jQuery.
+    resolve: {
+      alias: {
+        jquery: noop,
+        jsdom: noop
+      }
+    }
+  })
+}
+
+module.exports = onCreateWebpackConfig
